backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
